+++ model_catalogs/__init__.py
@@ -29,6 +29,7 @@ from .utils import (  # noqa
     find_filelocs,
     get_fresh_parameter,
     is_fresh,
+    status,
 )